refactored httpStatus code usage
