+++ hedera-mirror-rest/constants.js
@@ -113,6 +113,32 @@ const tokenTypeFilter = {
   NON_FUNGIBLE_UNIQUE: 'non_fungible_unique',
 };
 
+class StatusCode {
+  constructor(code, message) {
+    this.code = code;
+    this.message = message;
+  }
+
+  isClientError() {
+    return this.code >= 400 && this.code < 500;
+  }
+
+  toString() {
+    return `${this.code} ${this.message}`;
+  }
+}
+
+const httpStatusCodes = {
+  BAD_GATEWAY: new StatusCode(502, 'Bad gateway'),
+  BAD_REQUEST: new StatusCode(400, 'Bad request'),
+  INTERNAL_ERROR: new StatusCode(500, 'Internal error'),
+  NO_CONTENT: new StatusCode(204, 'No content'),
+  NOT_FOUND: new StatusCode(404, 'Not found'),
+  OK: new StatusCode(200, 'OK'),
+  PARTIAL_CONTENT: new StatusCode(206, 'Partial mirror node'),
+  SERVICE_UNAVAILABLE: new StatusCode(503, 'Service unavailable'),
+};
+
 module.exports = {
   characterEncoding,
   cloudProviders,
@@ -121,6 +147,7 @@ module.exports = {
   defaultCloudProviderEndpoints,
   entityColumns,
   filterKeys,
+  httpStatusCodes,
   networks,
   orderFilterValues,
   recordStreamPrefix,
