+++ hedera-mirror-rest/server.js
@@ -43,7 +43,7 @@ const topicmessage = require('./topicmessage');
 const transactions = require('./transactions');
 const {isTestEnv} = require('./utils');
 const {DbError} = require('./errors/dbError');
-const {handleError} = require('./middleware/httpStatusHandler');
+const {handleError} = require('./middleware/httpErrorHandler');
 const {metricsHandler, recordIpAndEndpoint} = require('./middleware/metricsHandler');
 const {serveSwaggerDocs} = require('./middleware/openapiHandler');
 const {responseHandler} = require('./middleware/responseHandler');
